+++ cypress/e2e/edge-application.cy.js
@@ -1,43 +1,5 @@
-import generateUniqueName from '../support/utils'
-
-const selectors = {
-  login: {
-    emailInput: '[data-testid="signin-block__email-input"]',
-    nextButton: '[data-testid="signin-block__next-button"] > .p-button-label',
-    passwordInput: '[data-testid="signin-block__password-input"] > .p-inputtext',
-    signInButton: '[data-testid="signin-block__signin-button"] > .p-button-label'
-  },
-  sidebar: {
-    menuToggleButton: '[data-testid="sidebar-block__toggle-button"] > .p-button-icon',
-    edgeApplicationMenuItem:
-      '[data-testid="sidebar-block__menu-item__edge-application"] > .p-menuitem-text'
-  },
-  edgeApplication: {
-    createButton: '.p-datatable-header > .flex-wrap > .p-button > .p-button-label',
-    nameInput: '[data-testid="form-horizontal-general-name__input"]',
-    addressInput: '[data-testid="form-horizontal-default-origin-address-field-text__input"]',
-    saveButton: '[data-testid="form-actions-submit-button"]',
-    cancelButton: '[data-testid="form-actions-cancel-button"]',
-    searchInput: '[data-testid="data-table-search-input"]',
-    tableRowName: '[data-testid="list-table-block__column__name__row"]',
-    tableRowLastEditor: '[data-testid="list-table-block__column__lastEditor__row"]',
-    rulesEngineTab: 'li:nth-child(6)',
-    addRuleButton: '[data-testid="rules-engine-create-button"]',
-    ruleNameInput: '[data-testid="rule-form-general-name__input"]',
-    criteriaOperatorDropdown:
-      '[data-testid="rule-form-criteria-item-conditional-operator"] > .p-dropdown-trigger',
-    criteriaOperator: 'li[aria-label="is equal"]',
-    criteriaInputValue:
-      '[data-testid="rule-form-criteria-item-conditional-input-field-text__input"]',
-    behaviorsDropdown: '[data-testid="rule-form-behaviors-item-name"] > .p-dropdown-trigger',
-    behaviors: '#behaviors\\[0\\]\\.name_4',
-    ruleTable: '.p-datatable-tbody > tr > :nth-child(2) > div',
-    actionsButton: '[data-testid="data-table-actions-column-body-actions-menu-button"]',
-    deleteButton: '[data-testid="data-table-actions-column-body-actions-menu"]',
-    deleteInput: '[data-testid="delete-dialog-confirmation-input-field"]',
-    confirmDeleteButton: '[data-testid="delete-dialog-footer-delete-button"]'
-  }
-}
+import generateUniqueName from '../support/utils';
+import selectors from '../support/selectors';
 
 const edgeApplicationName = generateUniqueName('EdgeApp')
 const rulesEngineName = generateUniqueName('RulesEng')
@@ -46,13 +8,10 @@ describe('Edge Application', () => {
   beforeEach(() => {
     // Login
     cy.login()
+    cy.openProductThroughSidebar('edge-application')
   })
 
   it('Create and delete an edge application, and create a rule', () => {
-    // Navigate to Edge Applications
-    cy.get(selectors.sidebar.menuToggleButton).click()
-    cy.get(selectors.sidebar.edgeApplicationMenuItem).click()
-
     // Create an edge application
     cy.get(selectors.edgeApplication.createButton).click()
     cy.get(selectors.edgeApplication.nameInput).type(edgeApplicationName)
@@ -85,20 +44,11 @@ describe('Edge Application', () => {
 
     // Verify the rule was created
     cy.get(selectors.edgeApplication.searchInput).type(rulesEngineName)
-    cy.get(selectors.edgeApplication.ruleTable)
-      .should('be.visible')
-      .should('have.text', rulesEngineName)
+    cy.get(selectors.edgeApplication.ruleTable).should('be.visible').should('have.text', rulesEngineName)
+  })
 
+  afterEach(() => {
     // Delete the edge application
-    cy.visit('/edge-applications')
-    cy.get(selectors.edgeApplication.searchInput).clear()
-    cy.get(selectors.edgeApplication.searchInput).type(edgeApplicationName)
-    cy.get(selectors.edgeApplication.tableRowName)
-      .should('be.visible')
-      .should('have.text', edgeApplicationName)
-    cy.get(selectors.edgeApplication.actionsButton).click()
-    cy.get(selectors.edgeApplication.deleteButton).click()
-    cy.get(selectors.edgeApplication.deleteInput).type('delete')
-    cy.get(selectors.edgeApplication.confirmDeleteButton).click()
+    cy.deleteProduct(edgeApplicationName,'/edge-applications')
   })
 })
