+++ cypress/e2e/teams-permissions.cy.js
@@ -1,50 +1,17 @@
-import generateUniqueName from '../support/utils'
+import generateUniqueName from '../support/utils';
+import selectors from '../support/selectors';
 
-const selectors = {
-  list: {
-    createTeamButton: '[data-testid="create_Team_button"]',
-    searchField: '[data-testid="data-table-search-input"]',
-    filteredRow: {
-      nameColumn: '.p-selectable-row > :nth-child(1)',
-      permissionsColumn: '.p-selectable-row > :nth-child(2)',
-      statusColumn: '.p-selectable-row > :nth-child(3)',
-      empty: 'tr.p-datatable-emptymessage > td'
-    },
-    actionsMenu: {
-      button: '[data-testid="data-table-actions-column-body-actions-menu-button"]',
-      deleteAction: '.p-menuitem-content > .p-menuitem-link'
-    },
-    deleteDialog: {
-      confirmationInputField: '[data-testid="delete-dialog-confirmation-input-field"]',
-      deleteButton: '[data-testid="delete-dialog-footer-delete-button"]'
-    }
-  },
-  form: {
-    teamName: '[data-testid="teams-permissions-form__name__field-text__input"]',
-    teamNameError: '[data-testid="teams-permissions-form__name__field-text__error-message"]',
-    teamStatus: '[data-testid="teams-permissions-form__form-fields__status"]',
-    allPermissionsToTarget: '[aria-label="Move All to Target"] > .p-icon',
-    allPermissionsToSource: '[aria-label="Move All to Source"] > .p-icon',
-    singlePermissionToTarget: 'button[aria-label="Move to Target"]',
-    viewContentDeliverySettingsPermission:
-      '[data-testid="teams-permissions-form__permissions-field__picklist__item-View Content Delivery Settings"]',
-
-    actionsSubmitButton: '[data-testid="form-actions-submit-button"]',
-    actionsCancelButton: '[data-testid="form-actions-cancel-button"]'
-  }
-}
-
-const teamName = generateUniqueName('Team')
+const teamsPermissionsName = generateUniqueName('TeamName')
 
 describe('Teams Permissions', () => {
   beforeEach(() => {
     cy.login()
-    cy.openMenuItem('Teams Permissions')
+    cy.openItemThroughMenuAccount('Teams Permissions');
   })
 
   it('should create a new team', () => {
     // Arrange
-    cy.get(selectors.list.searchField).type(teamName)
+    cy.get(selectors.list.searchField).type(teamsPermissionsName)
     cy.get(selectors.list.filteredRow.empty)
       .should('be.visible')
       .and('have.text', 'No teams found.')
@@ -59,7 +26,7 @@ describe('Teams Permissions', () => {
 
     // Act
     // Test different permissions scenarios and verify whether form is enabled/disabled
-    cy.get(selectors.form.teamName).type(teamName)
+    cy.get(selectors.form.teamName).type(teamsPermissionsName)
     cy.get(selectors.form.allPermissionsToTarget).click()
     cy.get(selectors.form.actionsSubmitButton).should('be.enabled')
 
@@ -73,29 +40,23 @@ describe('Teams Permissions', () => {
     cy.get(selectors.form.actionsSubmitButton).click()
 
     // Assert
-    cy.verifyToast('success', 'Your Team Permission has been created')
+    cy.get(selectors.toast.content)
+      .should('be.visible')
+      .and('contain', 'successYour Team Permission has been created')
+
     cy.get(selectors.form.actionsCancelButton).click()
 
-    cy.get(selectors.list.searchField).type(teamName)
-    cy.get(selectors.list.filteredRow.nameColumn).should('have.text', teamName)
+    cy.get(selectors.list.searchField).type(teamsPermissionsName)
+    cy.get(selectors.list.filteredRow.nameColumn).should('have.text', teamsPermissionsName)
     cy.get(selectors.list.filteredRow.permissionsColumn).should(
       'have.text',
       'View Content Delivery SettingsEdit Content Delivery SettingsShow more (46)'
     )
     cy.get(selectors.list.filteredRow.statusColumn).should('have.text', 'Active')
+  })
 
-    // Cleanup
-    cy.get(selectors.list.actionsMenu.button).click()
-    cy.get(selectors.list.actionsMenu.deleteAction).click()
-
-    cy.get(selectors.list.deleteDialog.confirmationInputField).type('delete')
-    cy.get(selectors.list.deleteDialog.deleteButton).click()
-
-    cy.get(selectors.list.searchField).clear()
-    cy.get(selectors.list.searchField).type(teamName)
-    cy.get(selectors.list.filteredRow.empty)
-      .should('be.visible')
-      .and('have.text', 'No teams found.')
-    cy.verifyToast('Team Permission successfully deleted')
+  afterEach(() => {
+    // Delete the team permissions
+    cy.deleteProduct(teamsPermissionsName, '/teams-permission')
   })
 })
