+++ cypress/e2mock/your-settings.cy.js
@@ -1,3 +1,5 @@
+import selectors from '../support/selectors';
+
 const payload = {
   country_call_code: 'BR - 55',
   email: 'testuser@cy.com',
@@ -11,43 +13,12 @@ const payload = {
   two_factor_enabled: true
 }
 
-const selectors = {
-  profile: {
-    firstNameInput: '[data-testid="profile__first-name__input"]',
-    firstNameError: '[data-testid="profile__first-name__error-message"]',
-    lastNameInput: '[data-testid="profile__last-name__input"]',
-    lastNameError: '[data-testid="profile__last-name__error-message"]',
-    timezoneOptions: '[data-testid="profile__timezone__options"]',
-    language: '[data-testid="profile__language"]'
-  },
-  contact: {
-    emailInput: '[data-testid="contact__email__input"]',
-    emailError: '[data-testid="contact__email__error-message"]',
-    mobileCountryCodeOptions: '[data-testid="contact__mobile__country-code-options"]',
-    countryCodeFilter: '.p-dropdown-filter',
-    countryCodeOption: (countryCode) => `#countryCallCode_${countryCode}`,
-    mobileInput: '[data-testid="contact__mobile__input"]',
-    mobileError: '[data-testid="contact__mobile__error-text"]'
-  },
-  security: {
-    oldPasswordInput: '[data-testid="security__old-password__input"]',
-    newPasswordInput: '[data-testid="security__new-password__input"]',
-    confirmPasswordInput: '[data-testid="security__confirm-password__input"]',
-    confirmPasswordError: '[data-testid="security__confirm-password__error-text"]',
-    twoFactorToggle: '.p-inputswitch-slider'
-  },
-  form: {
-    submitButton: '[data-testid="form-actions-submit-button"]'
-  }
-}
-
 describe('Your Settings spec', () => {
   beforeEach(() => {
     cy.login()
     cy.intercept('GET', '/api/v4/iam/user', { fixture: 'your-settings.json' }).as('getUser')
     cy.intercept('PATCH', '/api/v4/iam/user', { statusCode: 200, body: {} }).as('patchUser')
-
-    cy.openMenuItem('Your Settings')
+    cy.openItemThroughMenuAccount('Your Settings')
   })
 
   it('should edit user settings but the email', () => {
