+++ cypress/support/commands.js
@@ -1,31 +1,7 @@
 /* eslint-disable no-undef */
 /* eslint-disable no-console */
+import selectors from '../support/selectors';
 
-// Define selectors for elements
-const selectors = {
-  menu: {
-    avatarIcon: '[data-testid="profile-block__avatar"]',
-    menuItem: (menuItemLabel) =>
-      `li[aria-label="${menuItemLabel}"] > .p-menuitem-content > .p-menuitem-link`
-  },
-  login: {
-    emailInput: '[data-testid="signin-block__email-input"]',
-    nextButton: '[data-testid="signin-block__next-button"] > .p-button-label',
-    passwordInput: '[data-testid="signin-block__password-input"] > .p-inputtext',
-    signInButton: '[data-testid="signin-block__signin-button"] > .p-button-label'
-  },
-  sidebar: {
-    toggleButton: '[data-testid="sidebar-block__toggle-button"]',
-    menuItem: (productName) => `[data-testid="sidebar-block__menu-item__${productName}"]`
-  }
-}
-
-// Disable test failure for all uncaught exceptions
-Cypress.on('uncaught:exception', (err, runnable) => {
-  console.log('Uncaught exception in test:', runnable.title)
-  console.error('Uncaught exception:', err)
-  return false
-})
 
 // Function to perform login
 const login = (email, password) => {
@@ -35,6 +11,25 @@ const login = (email, password) => {
   cy.get(selectors.login.passwordInput).type(password, { log: false })
   cy.get(selectors.login.signInButton).click()
 }
+
+// Function to perform delete a product created
+const deleteProduct = (productName, path) => {
+  cy.visit(`${path}`);
+  cy.get(selectors.list.searchInput).clear()
+  cy.get(selectors.list.searchInput).type(productName)
+  cy.get(selectors.list.filteredRow.nameColumn).should('be.visible').should('have.text', productName)
+  cy.get(selectors.list.actionsMenu.button).click()
+  cy.get(selectors.list.actionsMenu.deleteButton).click()
+  cy.get(selectors.list.deleteDialog.confirmationInputField).type('delete')
+  cy.get(selectors.list.deleteDialog.deleteButton).click()
+};
+
+// Disable test failure for all uncaught exceptions
+Cypress.on('uncaught:exception', (err, runnable) => {
+  console.log('Uncaught exception in test:', runnable.title);
+  console.error('Uncaught exception:', err);
+  return false;
+});
 
 // Custom command to perform login using environment variables
 Cypress.Commands.add('login', () => {
@@ -46,15 +41,20 @@ Cypress.Commands.add('login', () => {
 
 // Custom command to open a product through the sidebar menu
 Cypress.Commands.add('openProductThroughSidebar', (productName) => {
-  cy.get(selectors.sidebar.toggleButton).click()
-  cy.get(selectors.sidebar.menuItem(productName)).click()
+  cy.get(selectors.menuSidebar.toggleButton).click()
+  cy.get(selectors.menuSidebar.menuItem(productName)).click()
 })
 
-// Custom command to open a menu item
-Cypress.Commands.add('openMenuItem', (menuItemLabel) => {
-  cy.get(selectors.menu.avatarIcon).click()
-  cy.get(selectors.menu.menuItem(menuItemLabel)).click()
-})
+// Custom command to open a item through the account menu
+Cypress.Commands.add('openItemThroughMenuAccount', (menuAccountLabel) => {
+  cy.get(selectors.menuAccount.avatarIcon).click();
+  cy.get(selectors.menuAccount.menuItem(menuAccountLabel)).click();
+});
+
+// Custom command to perform login using environment variables
+Cypress.Commands.add('deleteProduct', (productName, path) => {
+  deleteProduct(productName, path);
+});
 
 /**
  * Custom command to verify the visibility and content of a toast message.
